requirements.txt
streamlit
requests
python-dotenv
openai==1.86.0